fix(indexer): avoid long-context fp32 input expansion

+++ rtp_llm/models/deepseek_v2.py
@@ -234,7 +234,7 @@ class DeepSeekV2Weight(ModelDeployWeightInfo):
                             )
                         ],
                         transpose,
-                        data_type=torch.float32,
+                        data_type=torch.bfloat16,
                     ),
                 ]
             )
